Add files via upload
Explaining Bluetooth connections between an Android Smartphone and a robotic system built from scratch with an Arduino motherboard, a HC-06 slave Bluetooth module and a set of 4 LEDs placed forming a cross. The Android APP used is called "Bluetooth RC Car" and each of the 4 arrows (Forward, Backwards, Left or Right) will switch on the corresponding LED placed at the same position as the arrow when pressed.